Serve Swagger 2.0 API docs on /api-docs using Paperclip
This is far from perfect, but it works!
- This is version Swagger 2.0 (current is OpenAPI 3.0)
  - Paperclip dev says they would work on v3.0 for next release https://github.com/actix/actix-web/issues/310#issuecomment-628440449
- It doesn't seem to show descriptions of API operations (sigh), but it does
  show descriptions of *some* output schema fields.
- It even lists possible error codes and descriptions, but doesn't mention
  schema of the error response (undestandably).
- No Swagger UI is exposed, but there is an open MR for that in Paperclip:
  https://github.com/wafflespeanut/paperclip/pull/174
- `language` enum varians are nicely listed and correctly lowercased.

Given that Paperclip says it is WIP, this seems rather fine.

They also claim (about the actix support) that *While it's not feature complete,
you can rely on it to not break your actix-web flow.* And diff of this commit
confirms that, is is basically just addition of some derives, attributes and
wrappers.